% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metalite.table1-package.R
\docType{package}
\name{metalite.table1-package}
\alias{metalite.table1}
\alias{metalite.table1-package}
\title{metalite.table1: Interactive Table 1}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Yilong Zhang \email{elong0527@gmail.com}

}
\keyword{internal}
