% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metalite_table1.R
\name{metalite_table1}
\alias{metalite_table1}
\title{Interactive table1}
\usage{
metalite_table1(formula, data, id)
}
\arguments{
\item{formula}{a formula}

\item{data}{a data frame}

\item{id}{a variable name}
}
\description{
Interactive table1
}
