% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metalite_table1.R
\name{metalite_table1_to_df}
\alias{metalite_table1_to_df}
\title{Convert to data frame}
\usage{
metalite_table1_to_df(x)
}
\arguments{
\item{x}{an output from \code{metalite_table1}.}
}
\value{
a data frame that contain the table from the html output and
a \code{.id} variable for variables.
}
\description{
Convert to data frame
}
