% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metalite_table1.R
\name{metalite_table1_to_html}
\alias{metalite_table1_to_html}
\title{Convert to html}
\usage{
metalite_table1_to_html(x)
}
\arguments{
\item{x}{an output from \code{metalite_table1}.}
}
\value{
HTML string of \code{reactable} HTML widget for
interactive table of describptive statistics.
}
\description{
Convert to html
}
