% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metalite_table1_to_rtf.R
\name{metalite_table1_to_rtf}
\alias{metalite_table1_to_rtf}
\title{Convert \code{metalite_table1} to an RTF file}
\usage{
metalite_table1_to_rtf(
  x,
  file,
  title = "Baseline Characteristics",
  rel_width = NULL
)
}
\arguments{
\item{x}{Output of \code{metalite_table1}.}

\item{file}{A character string naming a file to save rtf file.}

\item{title}{Title in a character string.}
}
\value{
a string of the RTF file path
}
\description{
Convert \code{metalite_table1} to an RTF file
}
