% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactable2.R
\name{reactable_to_df}
\alias{reactable_to_df}
\title{Convert reactable to a data frame}
\usage{
reactable_to_df(x)
}
\arguments{
\item{x}{A \code{reactable} HTML widget}
}
\value{
A data frame
}
\description{
Convert reactable to a data frame
}
