% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metalite.table1-package.R
\docType{package}
\name{metalite.table1-package}
\alias{metalite.table1}
\alias{metalite.table1-package}
\title{metalite.table1: Interactive Table of Descriptive Statistics in HTML}
\description{
Create an interactive table of descriptive statistics in HTML. This table is typically used for exploratory analysis in a clinical study (referred to as 'Table 1').
}
\author{
\strong{Maintainer}: Yilong Zhang \email{elong0527@gmail.com}

}
\keyword{internal}
